+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Added support for the Librem Key
+- Added support for user provided extensions through lookup via the
+  `PATH` environment variable
 - Added the `fill` command that fills the SD card of a Nitrokey Storage device
   with random data
   - Added the `termion` dependency in version `1.5.5`